+++ R/model_parameters.other.R
@@ -140,3 +140,23 @@ model_parameters.coeftest <- function(model,
   attr(out, "object_name") <- deparse(substitute(model), width.cutoff=500)
   out
 }
+
+
+
+# ivprobit -------------------------------------------------------
+
+#' @export
+model_parameters.ivprobit <- function(model,
+                                      ci = .95,
+                                      verbose = TRUE,
+                                      ...) {
+  out <- .model_parameters_generic(
+    model = model,
+    ci = ci,
+    merge_by = "Parameter",
+    verbose = verbose,
+    ...
+  )
+  attr(out, "object_name") <- deparse(substitute(model), width.cutoff=500)
+  out
+}